backend/requirements.txt
accelerate==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
bitsandbytes==0.41.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
datasets==2.15.0
dill==0.3.7
fastapi==0.104.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2023.10.0
h11==0.16.0
hf-xet==1.2.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
networkx==3.6
numpy==1.24.3
packaging==25.0
pandas==2.3.3
peft==0.7.0
propcache==0.4.1
protobuf==4.25.1
psutil==7.1.3
pyarrow==22.0.0
pyarrow-hotfix==0.7
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
regex==2023.10.3
requests==2.32.5
safetensors==0.7.0
scipy==1.11.4
sentencepiece==0.1.99
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
